Re-organized interfaces.
