+++ src/backend/KnowledgeNetwork.Domains.Code/Analyzers/Classes/Abstractions/IInheritanceAnalyzer.cs
@@ -2,15 +2,15 @@ using KnowledgeNetwork.Domains.Code.Models.Classes;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-namespace KnowledgeNetwork.Domains.Code.Analyzers.Classes.RelationshipAnalyzers;
+namespace KnowledgeNetwork.Domains.Code.Analyzers.Classes.Abstractions;
 
 /// <summary>
-/// Analyzes dependency relationships (usage without ownership)
+/// Analyzes inheritance relationships between classes
 /// </summary>
-public interface IDependencyAnalyzer
+public interface IInheritanceAnalyzer
 {
     /// <summary>
-    /// Analyzes dependency relationships within the provided type declarations
+    /// Analyzes inheritance relationships within the provided type declarations
     /// </summary>
     Task AnalyzeAsync(
         SemanticModel semanticModel,
