+++ src/backend/KnowledgeNetwork.Domains.Code/Analyzers/Classes/Abstractions/INestedClassAnalyzer.cs
@@ -2,15 +2,15 @@ using KnowledgeNetwork.Domains.Code.Models.Classes;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-namespace KnowledgeNetwork.Domains.Code.Analyzers.Classes.RelationshipAnalyzers;
+namespace KnowledgeNetwork.Domains.Code.Analyzers.Classes.Abstractions;
 
 /// <summary>
-/// Analyzes interface implementation relationships
+/// Analyzes nested class relationships
 /// </summary>
-public interface IInterfaceImplementationAnalyzer
+public interface INestedClassAnalyzer
 {
     /// <summary>
-    /// Analyzes interface implementations within the provided type declarations
+    /// Analyzes nested class relationships within the provided type declarations
     /// </summary>
     Task AnalyzeAsync(
         SemanticModel semanticModel,
