+++ src/backend/KnowledgeNetwork.Domains.Code/Analyzers/Classes/RelationshipAnalyzers/InterfaceImplementationAnalyzer.cs
@@ -1,3 +1,4 @@
+using KnowledgeNetwork.Domains.Code.Analyzers.Classes.Abstractions;
 using KnowledgeNetwork.Domains.Code.Analyzers.Classes.Utilities;
 using KnowledgeNetwork.Domains.Code.Models.Classes;
 using Microsoft.CodeAnalysis;
