+++ src/backend/KnowledgeNetwork.Domains.Code/Analyzers/Classes/Utilities/ClassNodeFactory.cs
@@ -1,3 +1,4 @@
+using KnowledgeNetwork.Domains.Code.Analyzers.Classes.Abstractions;
 using KnowledgeNetwork.Domains.Code.Models.Classes;
 using KnowledgeNetwork.Domains.Code.Models.Common;
 using Microsoft.CodeAnalysis;
