+++ src/backend/KnowledgeNetwork.Domains.Code/Analyzers/Classes/Utilities/FileNameResolver.cs
@@ -1,3 +1,4 @@
+using KnowledgeNetwork.Domains.Code.Analyzers.Classes.Abstractions;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.Extensions.Logging;
